Render table only when drawer is open
